+++ src/poj_2538_47percentAC.cpp
@@ -6,7 +6,7 @@ int main() {
 	string str = "`1234567890-=QWERTYUIOP[]\ASDFGHJKL;'ZXCVBNM,./";
 	string input;
 	while (getline(cin,input)) {
-		for (int i = 0; i < input.length(); i++) {
+		for (size_t i = 0; i < input.length(); i++) {
 			if (input[i] == ' ')
 				continue;
 			input[i] = str[str.find(input[i]) - 1];
